requirements.txt
fastapi
uvicorn
pydantic
pytest
pytest-html
pytest-cov
httpx
python-jose
ruff
mypy

playwright
pytest-playwright